backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
crewai>=0.80.0
crewai-tools>=0.17.0
anthropic>=0.40.0
openai>=1.58.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
pydantic>=2.10.0
pillow>=10.4.0
moviepy>=1.0.3
aiofiles>=24.1.0
